test: invalid custom name
