--- tests/integration/custom_name/test_invalid_custom_name.py
@@ -1,22 +1,0 @@
-import os
-
-from jina import Client, Document, DocumentArray
-
-from jcloud.flow import CloudFlow
-
-flows_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'flows')
-flow_file = 'invalid-name.yml'
-protocol = 'grpc'
-
-
-def test_invalid_custom_name():
-    with CloudFlow(path=os.path.join(flows_dir, flow_file)) as flow:
-        assert flow.endpoints != {}
-        assert flow.endpoints['gateway'].startswith(f'{protocol}s://')
-        gateway = flow.endpoints['gateway']
-
-        da = Client(host=gateway).post(
-            on='/',
-            inputs=DocumentArray(Document(text=f'text-{i}') for i in range(50)),
-        )
-        assert len(da.texts) == 50
